requirements.txt
exchangelib==5.1.0
schedule==1.2.1
tk==0.1.0
SQLAlchemy==2.0.25
python-dotenv==1.0.1